Restructured the constants folder
