+++ src/components/Dialog/DrawAcceptDialog.js
@@ -2,7 +2,7 @@ import React from "react";
 import { useDispatch, useSelector } from "react-redux";
 import { Button, Dialog, DialogActions, DialogContent, DialogTitle } from "@mui/material";
 import { wsMssgDraw } from "../../actions/serverActions";
-import drawAcceptDialogActionTypes from "../../constants/drawAcceptDialogActionTypes";
+import drawAcceptDialogActionTypes from "../../constants/dialog/drawAcceptDialogActionTypes";
 import Wording from "../../utils/Wording.js";
 
 const DrawAcceptDialog = () => {
