+++ src/components/Dialog/FenDialog.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { Button, Dialog, DialogActions, DialogContent, DialogTitle, TextField } from '@mui/material';
 import { useDispatch, useSelector } from "react-redux";
-import fenDialogActions from '../../constants/fenDialogActionTypes';
+import fenDialogActions from '../../constants/dialog/fenDialogActionTypes';
 
 const FenDialog = () => {
   const state = useSelector(state => state);
