+++ src/constants/dialog/drawAcceptDialogActionTypes.js
