+++ src/constants/dialog/heuristicPictureDialogActionTypes.js
