+++ src/constants/dialog/openingSearchEcoDialogActionTypes.js
