+++ src/constants/dialog/openingSearchNameDialogActionTypes.js
