+++ src/constants/dialog/pgnDialogActionTypes.js
