+++ src/constants/dialog/resignAcceptDialogActionTypes.js
