+++ src/reducers/playLikeGrandmasterDialogReducer.js
@@ -1,4 +1,4 @@
-import playLikeGrandmasterDialogActionTypes from '../constants/playLikeGrandmasterDialogActionTypes';
+import playLikeGrandmasterDialogActionTypes from '../constants/dialog/playLikeGrandmasterDialogActionTypes';
 
 const initialState = {
   open: false
